\name{TCGAbrcaPackage}
\alias{TCGAbrcaPackage}
\docType{package}
\title{TCGAbrcaPackage:  TCGA Breast Invasive Adenocarcinoma data.
}
\description{
Acquired 2-10-15
}
\details{
\tabular{ll}{
Package: \tab TCGAbrca\cr
Type: \tab Package\cr
Version: \tab 0.99.9\cr
Date: \tab 2010-04-06\cr
License: \tab LGPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Paul Shannon <pshannon@systemsbiology.org>
Maintainer: Paul Shannon <pshannon@systemsbiology.org>
}
\references{
none yet...
}
\keyword{dynamic}
\seealso{
  NA
}
\examples{
  NA
}
